requirements.txt
Flask~=3.1.0
rich~=13.9.4
python-dotenv~=1.0.1
pymssql~=2.3.2
httpx[http2]~=0.28.1
celery~=5.4.0
redis~=5.0.0